Allow only the host to delete and edit an event.

+++ frontend/src/pages/Events/Events.jsx
@@ -8,7 +8,7 @@ import DisplayGrid from '../../components/DisplayGrid/DisplayGrid.jsx';
 // Libraries
 
 // Functions
-import {useEffect, useRef, useState} from 'react';
+import {useContext, useEffect, useRef, useState} from 'react';
 import {handleFormChange} from '../../helpers/handlers.js';
 import axios from 'axios';
 import {API} from '../../globalConstants.js';
@@ -20,68 +20,43 @@ import DatePanel from 'react-multi-date-picker/plugins/date_panel';
 import SearchBar from '../../components/SearchBar/SearchBar.jsx';
 import {useNavigate} from 'react-router-dom';
 import EventForm from '../../components/EventForm/EventForm.jsx';
+import {AuthContext} from '../../context/AuthContext.jsx';
 
 function Events() {
     //<editor-fold desc="State">
-    const initialEventFormState = {
-        name: '',
-        location: '',
-        gameId: 0,
-        isHostPlaying: false,
-        definitiveTime: '',
-        possibleTimes: [],
-    }
     const [param, setParam] = useState('');
     const [eventId, setEventId] = useState(2);
     const [updated, setUpdated] = useState(0);
     //</editor-fold>
 
     const navigate = useNavigate();
-
-    //<editor-fold desc="Effects">
-    //</editor-fold>
+    const {user, isAdmin, isUser} = useContext(AuthContext);
 
     //<editor-fold desc="Handlers">
-    function handleDeleteChange(e) {
-        const newValue = e.target.value;
-
-        if (newValue <= 1) {
-            setEventId(2);
+    function createEvent(e) {
+        if (isUser) {
+            navigate('/events/create')
         } else {
-            setEventId(newValue);
+            console.error('Only a user can create an event.')
         }
     }
-
-    async function handleDeleteEventSubmit(e) {
-        e.preventDefault();
-        try {
-            const response = await axios.delete(`${API}/events/${eventId}`, {
-                headers: {
-                    Authorization: `Bearer ${localStorage.getItem('token')}`,
-                }
-            });
-            console.log(response);
-        } catch (er) {
-            console.error(er.message + ': ' + er.response.data);
-            console.error(`${API}/events/${eventId}`);
-        }
-        await new Promise(r => setTimeout(r, 2000)); // small delay
-        setUpdated(updated+1);
-    }
-
     async function deleteEvents(e) {
         e.preventDefault();
-        try {
-            const response = await axios.delete(`${API}/events/deleteAll`, {
-                headers: {
-                    Authorization: `Bearer ${localStorage.getItem('token')}`,
-                },
-            });
-            console.log(response);
-        } catch (er) {
-            console.error(er);
+        if (isAdmin) {
+            try {
+                const response = await axios.delete(`${API}/events/deleteAll`, {
+                    headers: {
+                        Authorization: `Bearer ${localStorage.getItem('token')}`,
+                    },
+                });
+                console.log(response);
+            } catch (er) {
+                console.error(er);
+            }
+            setUpdated(updated + 1);
+        } else {
+            console.error('Only an admin can delete all events.')
         }
-        setUpdated(updated+1);
     }
     //</editor-fold>
 
@@ -89,21 +64,7 @@ function Events() {
         <div className="categoryPage">
             <h2>Events</h2>
 
-            <button type="button" onClick={() => navigate('/events/create')}>Create Event</button>
-
-            {/*<editor-fold desc="Delete Event Form">*/}
-            <form onSubmit={handleDeleteEventSubmit}>
-                <label htmlFor="deleteEventId">Event ID:</label>
-                <input
-                    type="number"
-                    name="deleteEventId"
-                    id="deleteEventId"
-                    value={eventId}
-                    onChange={handleDeleteChange}
-                />
-                <button type="submit">Delete event</button>
-            </form>
-            {/*</editor-fold>*/}
+            <button type="button" onClick={createEvent}>Create Event</button>
 
             <button type="button" onClick={deleteEvents}>Delete all</button>
